Initial commit
Note that the original repo was shallow cloned with --depth=1 and then I
added a single commit on top.

unfortunately in order to push this to a new repo I've had to unshallow
the clone - I opted to not fetch the whole of Sandi's history and
decided that I would rather just scorch earth on the whole history.

So this commit is the example repo - with my 30 minute timeboxed
solution already in place